refactor: move middlewares to their own respective packages
