+++ v2/middlewares/backoff/backoff_test.go
@@ -1,4 +1,4 @@
-package middlewares_test
+package backoff_test
 
 import (
 	"context"
@@ -6,24 +6,25 @@ import (
 	"testing"
 	"time"
 
+	"github.com/honestbank/kp/v2/middlewares/backoff"
+
 	"github.com/confluentinc/confluent-kafka-go/kafka"
 	"github.com/stretchr/testify/assert"
 
 	backoff_policy "github.com/honestbank/backoff-policy"
-	"github.com/honestbank/kp/v2/middlewares"
 )
 
 func TestBackoff(t *testing.T) {
 	t.Run("calls next", func(t *testing.T) {
 		called := false
-		middlewares.Backoff(backoff_policy.NewExponentialBackoffPolicy(0, 0)).Process(context.Background(), nil, func(ctx context.Context, msg *kafka.Message) error {
+		backoff.NewBackoffMiddleware(backoff_policy.NewExponentialBackoffPolicy(0, 0)).Process(context.Background(), nil, func(ctx context.Context, msg *kafka.Message) error {
 			called = true
 			return nil
 		})
 		assert.True(t, called)
 	})
 	t.Run("returns what next returns", func(t *testing.T) {
-		mw := middlewares.Backoff(backoff_policy.NewExponentialBackoffPolicy(0, 0))
+		mw := backoff.NewBackoffMiddleware(backoff_policy.NewExponentialBackoffPolicy(0, 0))
 		err := errors.New("some error")
 		actualErr := mw.Process(context.Background(), nil, func(ctx context.Context, msg *kafka.Message) error {
 			return err
@@ -31,7 +32,7 @@ func TestBackoff(t *testing.T) {
 		assert.Same(t, err, actualErr)
 	})
 	t.Run("when there's error, it slows down", func(t *testing.T) {
-		mw := middlewares.Backoff(backoff_policy.NewExponentialBackoffPolicy(time.Second, 5))
+		mw := backoff.NewBackoffMiddleware(backoff_policy.NewExponentialBackoffPolicy(time.Second, 5))
 		_ = mw.Process(context.Background(), nil, func(ctx context.Context, msg *kafka.Message) error {
 			return errors.New("some error")
 		})
